requirements.txt
Flask==3.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pypdf==3.17.4
python-dotenv==1.0.0
werkzeug==3.0.1
pandas==2.1.4
tabulate==0.9.0
zipfile36==0.1.3
python-magic==0.4.27
requests==2.31.0
markupsafe==2.1.3
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
